feat: S3 scan object assume role in account
Update the S3 scan object function to assume a role in
the target account to perform S3 object tagging.  A new S3Client
will be created and cached for each target account ID in the
event payload.

Update the Scan Files API to add the AWS account ID to the SNS
message published to the S3 scan object function.