fix: use default SecretsManager encryption key
Update the Scan Files API auth token secret to use the default
SecretsManager encryption key.

The reason we had initially used a CMK was because we were
providing cross-account access to the secret, so a CMK was
required for the decryption.  Now that cross-account access
secret access has been removed, we no longer need to use
the CMK.