fix: `ERROR` logs from ClamAV daemon startup
Update the logging from the ClamAV daemon startup to log less errors
during the daemon startup.  This is being done so that we can setup
CloudWatch alarms that trigger on unexpected API errors.

Part of this change is to add more log output around failures to get
the daemon's PID so that we can tune the exception handling in the
`get_clamd_pid` method.