fix: full terraform apply in Staging (#192)
* fix: full terraform apply in Staging

Update the Staging `terraform apply` to always run against
all modules.  The reason for this change is because it was allowing
for scenarios where ClickOps changes were not getting reset
since there were no changes to the affected modules.

Although this increases blast radius potential for changes,
it does reduce the potential for drift and matches the Staging
Terraform plan/apply results to how Production behaves.