fix: integration test and async event timeout (#214)
Update the timeout on the s3-scan-object lambda function so that it
can handle cases where the API needs to start the ClamAV daemon.

Perviously, the S3 event triggering the lambda function could timeout
if a new API function had just been deployed.  Since the S3 event
trigger is async, this would lead to multiple retries of the same
event, which would then fail once the workflow deleted the file.

To be on the safe side, we'll rely on S3 bucket expiration rules to
remove the integration test file so that mutliple retries of an S3
event don't lead to errors being logged.