feat: create API CloudWatch warning alarm
Add a CloudWatch alarm for API warnings and update the
`is_clamd_running` function to log a warning when it cannot
connect to the socket.

The reason for this is because returning `False` from this
function is an expected state, and we only want to catch cases
where it is continually occurring within a short period.