feat: allow S3 scan object read from SQS
Update the S3 scan object function to allow it to use the
SQS queue as an S3 event source.  Each account that uses
the S3 scan object module will need to have its account ID
added once it has created the source SQS queue.

Also includes a fix for the module code to JSON parse
the event string in the SQS message body.