fix: improve API secret retrieval validation
Update the API's entry script to more carefully check if the
secrets have been correctly retrieved.

Update the API's token verification middleware to warn when
an unauthorized request has been made.  This may end up
being noisier than we'd like, but it will help troubleshoot
the errors we've been seeing in the short term and can
be adjusted if needed.