fix: move API auth token into SSM ParameterStore
Update the SSM ParameterStore `ENVIRONMENT_VARIABLES` to
include the API auth token.  In testing it was discovered that the
SecretManager retrieval is behaving asynchronously so a small
percentage of Lambda initializations are missing the correct
value.

Update the check for an `API_AUTH_TOKEN` to the first
section of the API where it is initialized.