fix: remove in progress S3 objec tags (#1040)
Update the s3-scan-object module so that in progress S3 object tags
are no longer added.  These have been removed because in rare cases
we were seeing race conditions where the in progress tagging
operation overwrote the scan complete tags.

This occurs because S3 tag operations always replace all tags on an
object with the new TagSet in the request.